updated landing with code improvements
