+++ src/screens/HomeScreen.tsx
@@ -19,13 +19,7 @@ export const HomeScreen = () => {
       <Text.XxlBold textAlign="center">{t('home_screen.header_title')}</Text.XxlBold>
       <Text.LgRegular textAlign="center">{t('home_screen.header_subtitle')}</Text.LgRegular>
 
-      <Box
-        alignItems={'center'}
-        flexDirection={'row'}
-        flexWrap={'wrap'}
-        justifyContent={'center'}
-        mt={4}
-      >
+      <Box alignItems="center" flexDirection="row" flexWrap="wrap" justifyContent="center" mt={4}>
         <Button h={12} m={3} maxWidth={160} minWidth={160} onPress={() => openLink(BACA_DOCS_URL)}>
           {t('home_screen.read_docs')}
         </Button>
