+++ src/screens/index.ts
@@ -15,4 +15,4 @@ export * from './SettingsScreen'
 export * from './TestFormScreen'
 export * from './TypographyScreen'
 export * from './UserSessionScreen'
-export * from '../screens/drafts/images'
+export * from './drafts/images'
